Correctly override res.end
`res.end` is not an async function, therefore it cannot be overriden with an async function.
In that case its behavior change and the original `res.end` would be called too
late. There is no need to make this function async because what we want to do is to
log something at the end. And then, this function should return whatever the original one returns,
and it should pass the original paramters so this modification is done too.
The correct way I think is to use the `finish` event of the res object instead. I'll propose a PR for this too.
Meanwhile this close #12